requirements.txt
aiogram==3.13.1
python-docx==1.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
aiofiles==24.1.0
